rename ConfigSchema
