+++ mlos_bench/mlos_bench/optimizers/one_shot_optimizer.py
@@ -9,7 +9,7 @@ No-op optimizer for mlos_bench that proposes a single configuration.
 import logging
 from typing import Dict, Optional, Any
 
-from mlos_bench.config.schemas import ConfigSchemaType
+from mlos_bench.config.schemas import ConfigSchema
 from mlos_bench.services.base_service import Service
 from mlos_bench.tunables.tunable_groups import TunableGroups
 from mlos_bench.optimizers.mock_optimizer import MockOptimizer
@@ -36,7 +36,7 @@ class OneShotOptimizer(MockOptimizer):
             # Now assign the values we were given in the config.
             for data_file in config.get("include_tunable_values", []):
                 tunable_values = self._service.config_loader_service.load_config(data_file,
-                                                                                 schema_type=ConfigSchemaType.TUNABLE_VALUES)
+                                                                                 schema_type=ConfigSchema.TUNABLE_VALUES)
                 assert isinstance(tunable_values, Dict)
                 self._tunables.assign(tunable_values)
         self._tunables.assign(config.get("tunable_values", {}))
