+++ mlos_bench/mlos_bench/services/types/config_loader_type.py
@@ -8,7 +8,7 @@ Protocol interface for helper functions to lookup and load configs.
 
 from typing import List, Iterable, Optional, Union, Protocol, runtime_checkable, TYPE_CHECKING
 
-from mlos_bench.config.schemas import ConfigSchemaType
+from mlos_bench.config.schemas import ConfigSchema
 
 
 # Avoid's circular import issues.
@@ -43,7 +43,7 @@ class SupportsConfigLoading(Protocol):
             An actual path to the config or script.
         """
 
-    def load_config(self, json_file_name: str, schema_type: Optional[ConfigSchemaType]) -> Union[dict, List[dict]]:
+    def load_config(self, json_file_name: str, schema_type: Optional[ConfigSchema]) -> Union[dict, List[dict]]:
         """
         Load JSON config file. Search for a file relative to `_config_path`
         if the input path is not absolute.
@@ -53,7 +53,7 @@ class SupportsConfigLoading(Protocol):
         ----------
         json_file_name : str
             Path to the input config file.
-        schema_type : Optional[ConfigSchemaType]
+        schema_type : Optional[ConfigSchema]
             The schema type to validate the config against.
 
         Returns
