+++ mlos_bench/mlos_bench/tests/config/optimizers/test_load_optimizer_config_examples.py
@@ -15,7 +15,7 @@ import pytest
 
 from mlos_bench.tests.config import locate_config_examples
 
-from mlos_bench.config.schemas import ConfigSchemaType
+from mlos_bench.config.schemas import ConfigSchema
 from mlos_bench.services.config_persistence import ConfigPersistenceService
 from mlos_bench.optimizers.base_optimizer import Optimizer
 from mlos_bench.tunables.tunable_groups import TunableGroups
@@ -42,7 +42,7 @@ assert configs
 @pytest.mark.parametrize("config_path", configs)
 def test_load_optimizer_config_examples(config_loader_service: ConfigPersistenceService, config_path: str) -> None:
     """Tests loading a config example."""
-    config = config_loader_service.load_config(config_path, ConfigSchemaType.OPTIMIZER)
+    config = config_loader_service.load_config(config_path, ConfigSchema.OPTIMIZER)
     assert isinstance(config, dict)
     cls = get_class_from_name(config["class"])
     assert issubclass(cls, Optimizer)
