+++ mlos_bench/mlos_bench/tests/services/config_persistence_test.py
@@ -10,7 +10,7 @@ import os
 import sys
 import pytest
 
-from mlos_bench.config.schemas import ConfigSchemaType
+from mlos_bench.config.schemas import ConfigSchema
 from mlos_bench.services.config_persistence import ConfigPersistenceService
 
 
@@ -80,7 +80,7 @@ def test_load_config(config_persistence_service: ConfigPersistenceService) -> No
     Check if we can successfully load a config file located relative to `config_path`.
     """
     tunables_data = config_persistence_service.load_config("tunables/tunable-values-example.jsonc", schema_type=None)
-    # TODO: ConfigSchemaType.TUNABLE_VALUES)
+    # TODO: ConfigSchema.TUNABLE_VALUES)
     assert tunables_data is not None
     assert isinstance(tunables_data, dict)
     assert len(tunables_data) >= 1
